Refactoring the S2I command line interface
Closes #443
