+++ pkg/cmd/cmd.go
@@ -1,0 +1,1 @@
+package cmd
